Display All Data from JSON Files
Now outputs all json file data to the screen.
Takes the file directory, goes through all the files to only get JSON files and then addeds them to a list to be decoded for display.
This push will also include a few json data files that can be used from now on for testing.
Want to get the files to be properly combined, been looking into this but so far not managed to get it to work.

(Also updated(added) a git ignore for VS files.)