Started working through data inside JSON
Started working with the JS Object data taken from the json file. trying to get this to work without the object keys being hard coded but instead generated by the code. if this can be done the many different game data json files could be looked into rather than a only being able to look at a single game data type that is 100% tied to this.