Major Update.
Added proper combinging and saving of the json files, this saves a new json file to the folder "saved", this file is used for the rest of the functions that will read from it all the data in the file.
Added displaying all the data in that file properly so it can actually be read, reads out the saved combined json file and displays it in a way that can be read easily, just displaying the data nicely and not as if it was raw json data.
Added the ability to select a specific event type and view it all the entries of that event in the combined json file.