Added move visual features
Added a bar graph that allows users to see all the different events and the number of entries of that event.
the bar graph elements can be clicked on to display a pie chart with what session all that data is from. the same pie chart is also shown when using the option to display that specific data.